Move `find_path` to `utils`, to simplify
